chore: reorganize runtime tests.
In anticipation of the `thread_pool` module being switched to private,
tests are updated to use `Runtime` directly instead of `thread_pool`.
Those tests that cannot be updated are switched to unit tests.
